+++ libs/client/cornflow_client/schema/dict_functions.py
@@ -1,12 +1,11 @@
 """
 
 """
-# Partial imports
 
 from functools import partial
+
 from marshmallow import Schema, fields, INCLUDE
 
-# Imports from internal modules
 from cornflow_client.constants import (
     BOOLEAN_TYPE,
     STRING_TYPE,
@@ -73,6 +72,12 @@ def gen_schema(cls_name, params, possible_dict=None):
             p.pop("many", None)
         valid_values = p.pop("valid_values", None)
         name = p.pop("name")
+        # Move strict to metadata if it exists
+        strict = p.pop("strict", True)
+        metadata = p.pop("metadata", {})
+        metadata["strict"] = strict
+        p["metadata"] = metadata
+
         if valid_values is not None:
             dict_fields[name] = field_type(
                 validate=partial(validator, valid_values), **p